This is our project initial state.

+++ main.py
@@ -2,7 +2,7 @@ name = input("what is your is name? ")
 weight_kg = input("what is your weight in kg? ")
 height_m = input("what is your height in meters? ")
 height_m_sqr = float(height_m) * float(height_m)
-bmi = int(weight_kg) / float(height_m_sqr)
+bmi = float(weight_kg) / float(height_m_sqr)
 roundBMI = round(bmi, 1)
 print(roundBMI)
 if bmi < 25:
